Add new Files

+++ app/Http/Controllers/Admin/Roles/RolesController.php
@@ -17,8 +17,12 @@ class RolesController extends Controller
     public function index()
     {
         $roles = Role::all();
+        $permissions = Permission::all();
 
-        return view('admin.roles.index', ['roles' => $roles]);
+        return view('admin.roles.index', [
+            'roles' => $roles,
+            'permissions' => $permissions
+        ]);
     }
 
     /**
@@ -41,29 +45,35 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'name'=>'required|unique:roles|max:20',
-            'permissions' =>'required',
-            ]
-        );
+        if ($request->ajax()) {
+            $this->validate($request, [
+                'name'=>'required|unique:roles|max:20',
+                'permissions' =>'required',
+            ]);
 
-        $name = $request['name'];
-        $role = new Role();
-        $role->name = $name;
+            $name = $request['name'];
+            $role = new Role();
+            $role->name = $name;
 
-        $permissions = $request['permissions'];
+            $permissions = $request['permissions'];
 
-        $role->save();
-        //Looping thru selected permissions
-        foreach ($permissions as $permission) {
-            $p = Permission::where('id', '=', $permission)->firstOrFail(); 
-            //Fetch the newly created role and assign permission
-            $role = Role::where('name', '=', $name)->first(); 
-            $role->givePermissionTo($p);
+            $role->save();
+            //Looping thru selected permissions
+            foreach ($permissions as $permission) {
+                $p = Permission::where('id', '=', $permission)->firstOrFail(); 
+                //Fetch the newly created role and assign permission
+                $role = Role::where('name', '=', $name)->first(); 
+                $role->givePermissionTo($p);
+            }
+
+            return response([
+                'data' => [
+                    'role' => $role,
+                    'permissions' => str_replace(array('[',']','"'),' ', $role->permissions()->pluck('name'))
+                ],
+                'message' => 'success'
+            ]);
         }
-
-        return redirect()->route('roles.index')
-            ->with('flash_message', 'Role'. $role->name.' added!'); 
     }
 
     /**
@@ -85,10 +95,25 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $role = Role::findOrFail($id);
-        $permissions = Permission::all();
+        if (request()->ajax()) {
+            $role = Role::findOrFail($id);
+            $permissions = Permission::all();
 
-        return view('admin.roles.edit', ['role' => $role, 'permissions' => $permissions]);
+            if ($role) {
+                return response([
+                    'data' => [
+                        'role' => $role,
+                        'permissions' => $role->permissions->pluck('id')
+                    ],
+                    'message' => 'success'
+                ]);
+            }
+
+            return response([
+                'data' => '',
+                'message' => 'error'
+            ]);
+        }
     }
 
     /**
@@ -100,30 +125,39 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $role = Role::findOrFail($id);//Get role with the given id
-        //Validate name and permission fields
-        $this->validate($request, [
-            'name'=>'required|max:20|unique:roles,name,'.$id,
-            'permissions' =>'required',
-        ]);
+        if ($request->ajax()) {
+            $role = Role::findOrFail($id);//Get role with the given id
+            //Validate name and permission fields
+            $this->validate($request, [
+                'name'=>'required|max:20|unique:roles,name,'.$id,
+                'permissions' =>'required',
+            ]);
 
-        $input = $request->except(['permissions']);
-        $permissions = $request['permissions'];
-        $role->fill($input)->save();
+            $input = $request->except(['permissions']);
+            $permissions = $request['permissions'];
+            $role->fill($input)->save();
 
-        $p_all = Permission::all();//Get all permissions
+            $p_all = Permission::all();//Get all permissions
 
-        foreach ($p_all as $p) {
-            $role->revokePermissionTo($p); //Remove all permissions associated with role
+            foreach ($p_all as $p) {
+                $role->revokePermissionTo($p); //Remove all permissions associated with role
+            }
+
+            foreach ($permissions as $permission) {
+                $p = Permission::where('id', $permission)->firstOrFail(); //Get corresponding form //permission in db
+                $role->givePermissionTo($p->name);  //Assign permission to role
+            }
+
+            
+            return response([
+                'data' => [
+                    'role' => $role,
+                    'permissions' => str_replace(array('[',']','"'),' ', $role->permissions()->pluck('name'))
+                ],
+                'message' => 'Role Updated Successfully!'
+            ]);
+
         }
-
-        foreach ($permissions as $permission) {
-            $p = Permission::where('id', $permission)->firstOrFail(); //Get corresponding form //permission in db
-            $role->givePermissionTo($p->name);  //Assign permission to role
-        }
-
-        return redirect()->route('roles.index')
-            ->with('flash_message', 'Role'. $role->name.' updated!');
     }
 
     /**
@@ -134,10 +168,15 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
-        $role = Role::findOrFail($id);
-        $role->delete();
+        if (request()->ajax()) {
+            $role = Role::findOrFail($id);
+            $id = $role->id;
+            $role->delete();
 
-        return redirect()->route('roles.index')
-            ->with('flash_message', 'Role deleted!');
+            return response([
+                'data' => $id,
+                'message' => 'success'
+            ]);
+        }
     }
 }
